requirements.txt
anyio==4.8.0
APScheduler==3.11.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-telegram-bot==21.10
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
telegram==0.0.1
typing_extensions==4.12.2
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
